[AOT Refactor] subclass utils
[ghstack-poisoned]
